+++ test/unit/renderer/viz/expressions/aggregation/clusterAggregation.test.js
@@ -7,7 +7,8 @@ describe('src/renderer/viz/expressions/clusterAggregation', () => {
         _cdb_agg_min_price: 2,
         _cdb_agg_avg_price: 3,
         _cdb_agg_sum_price: 4,
-        _cdb_agg_mode_price: 5
+        _cdb_agg_mode_price: 5,
+        _cdb_feature_count: 1
     };
 
     let $price = null;
@@ -28,11 +29,13 @@ describe('src/renderer/viz/expressions/clusterAggregation', () => {
         validateMaxArgumentsError('clusterSum', ['number', 'number']);
         validateMaxArgumentsError('clusterAvg', ['number', 'number']);
         validateMaxArgumentsError('clusterMode', ['number', 'number']);
+        validateMaxArgumentsError('clusterCount', [0]);
     });
 
     describe('type', () => {
         validateStaticType('clusterMax', ['number-property'], 'number');
         validateStaticType('clusterMode', ['category-property'], 'category');
+        validateStaticType('clusterCount', [], 'number');
     });
 
     describe('eval', () => {
@@ -59,6 +62,16 @@ describe('src/renderer/viz/expressions/clusterAggregation', () => {
         it('clusterSum($price) should return fakeFeature._cdb_agg_sum_price', () => {
             const actual = s.clusterSum($price).eval(fakeFeature);
             expect(actual).toEqual(fakeFeature._cdb_agg_sum_price);
+        });
+
+        it('clusterCount() should return fakeFeature._cdb_feature_count if defined', () => {
+            const actual = s.clusterCount().eval(fakeFeature);
+            expect(actual).toEqual(fakeFeature._cdb_feature_count);
+        });
+
+        it('clusterCount() should return 1 if _cdb_feature_count is undefined in the feature', () => {
+            const actual = s.clusterCount().eval({});
+            expect(actual).toEqual(1);
         });
     });
 });