+++ CHANGELOG.md
@@ -5,7 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add new `clusterCount()` expression
 - Support feature-dependant arrays in combination with ramp
+
+### Fixed
+- Add default export to allow `import carto from '@carto/carto-vl';`
 
 ## [0.7.0] - 2018-08-24
 ### Added